requirements.txt
absl-py==2.1.0
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.4
contourpy==1.2.1
cycler==0.12.1
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
glog==0.3.1
grpcio==1.64.1
humanfriendly==10.0
importlib_metadata==8.0.0
importlib_resources==6.4.0
intel-openmp==2021.4.0
Jinja2==3.1.4
kiwisolver==1.4.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
onnx==1.16.0
onnxruntime==1.15.1
onnxruntime_extensions==0.11.0
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
protobuf==4.25.3
Pygments==2.18.0
pyparsing==3.1.2
pypiwin32==223
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
python-gflags==3.1.2
pyttsx3==2.90
pywin32==306
rich==13.7.1
six==1.16.0
sympy==1.12.1
tbb==2021.13.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
typing_extensions==4.12.27
Werkzeug==3.0.3
zipp==3.19.2
